Move compose variables to prebuilt-themes
