+++ room-kit/src/main/java/live/hms/roomkit/ui/meeting/activespeaker/HlsFragment.kt
@@ -131,12 +131,12 @@ import live.hms.roomkit.ui.meeting.chat.combined.ChatRbacRecipientHandling
 import live.hms.roomkit.ui.meeting.chat.combined.LaunchMessageOptionsDialog
 import live.hms.roomkit.ui.meeting.chat.combined.PinnedMessageUiUseCase
 import live.hms.roomkit.ui.meeting.chat.rbac.RoleBasedChatBottomSheet
-import live.hms.roomkit.ui.meeting.compose.Variables
-import live.hms.roomkit.ui.meeting.compose.Variables.Companion.OnSurfaceHigh
-import live.hms.roomkit.ui.meeting.compose.Variables.Companion.PrimaryDefault
-import live.hms.roomkit.ui.meeting.compose.Variables.Companion.Spacing0
-import live.hms.roomkit.ui.meeting.compose.Variables.Companion.Spacing1
-import live.hms.roomkit.ui.meeting.compose.Variables.Companion.Spacing2
+import live.hms.prebuilt_themes.Variables
+import live.hms.prebuilt_themes.Variables.Companion.OnSurfaceHigh
+import live.hms.prebuilt_themes.Variables.Companion.PrimaryDefault
+import live.hms.prebuilt_themes.Variables.Companion.Spacing0
+import live.hms.prebuilt_themes.Variables.Companion.Spacing1
+import live.hms.prebuilt_themes.Variables.Companion.Spacing2
 import live.hms.roomkit.ui.polls.leaderboard.millisToText
 import live.hms.roomkit.ui.theme.applyTheme
 import live.hms.roomkit.util.contextSafe
