+++ vb-prebuilt/src/main/java/live/hms/vb_prebuilt/VirtualBackgroundBottomSheet.kt
@@ -55,7 +55,7 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import live.hms.roomkit.R
-import live.hms.roomkit.ui.meeting.compose.Variables
+
 class VirtualBackgroundBottomSheet : BottomSheetDialogFragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
